ai_bot/requirements.txt
pynng
openai
trio
python-dotenv